refactor view helpers
