+++ src/data_access/APIDataAccess.java
@@ -29,7 +29,7 @@ public class APIDataAccess implements SingleStockAPIDataAccessInterface, SearchA
     }
 
     /**
-     * Set the HistoricalPrice and Exchange for the given stock using responses from the api call time series
+     * Set the HistoricalPrice for the given stock using responses from the api call time series
      * @param symbol the stock symbol required for the api call
      * @param interval the time interval required for the api call
      * @param outputSize the output size required for the api call
@@ -64,7 +64,7 @@ public class APIDataAccess implements SingleStockAPIDataAccessInterface, SearchA
     }
 
     /**
-     * Set the detailed info including stock name, currency, country and type for the given stock using response from
+     * Set the detailed info including stock name, exchange, currency, country and type for the given stock using response from
      * the api call stocks list.
      * @param symbol the stock symbol required for the api call
      * @param exchange the stock market exchange of the stock
