+++ src/interface_adapter/menu/MenuViewModel.java
@@ -8,7 +8,7 @@ import java.beans.PropertyChangeSupport;
 public class MenuViewModel extends ViewModel {
     public static final String SEARCH_BUTTON_LABEL = "Search";
     public static final String TITLE_LABEL = "Search Menu";
-    public static final String SEARCH_LABEL = "NASDAQ Stock Symbol";
+    public static final String SEARCH_LABEL = "Stock Symbol";
     private final PropertyChangeSupport support = new PropertyChangeSupport(this);
     private MenuState state = new MenuState();
