+++ src/view/helpers/LabelTextPanel.java
@@ -1,4 +1,4 @@
-package view;
+package view.helpers;
 
 import javax.swing.*;
